Extract component for showing lists of AppMediaItem
