+++ composeApp/src/commonMain/kotlin/io/music_assistant/client/ui/compose/library/LibraryListScreen.kt
@@ -3,43 +3,32 @@
 package io.music_assistant.client.ui.compose.library
 
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.grid.rememberLazyGridState
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.automirrored.filled.ViewList
 import androidx.compose.material.icons.filled.GridView
 import androidx.compose.material.icons.filled.Search
 import androidx.compose.material.icons.filled.SearchOff
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import compose.icons.TablerIcons
-import compose.icons.tablericons.Plus
 import io.music_assistant.client.data.model.client.ClickContext
 import io.music_assistant.client.data.model.client.LibraryFilters
 import io.music_assistant.client.data.model.client.MediaType
@@ -51,13 +40,6 @@ import io.music_assistant.client.ui.compose.common.DataState
 import io.music_assistant.client.ui.compose.common.SelectOption
 import io.music_assistant.client.ui.compose.common.SortChip
 import io.music_assistant.client.ui.compose.common.ToastHost
-import io.music_assistant.client.ui.compose.common.ToastState
-import io.music_assistant.client.ui.compose.common.clearFocusOnScroll
-import io.music_assistant.client.ui.compose.common.items.CreatePlaylistDialog
-import io.music_assistant.client.ui.compose.common.items.LibraryActions
-import io.music_assistant.client.ui.compose.common.items.PlayHandler
-import io.music_assistant.client.ui.compose.common.items.PlaylistActions
-import io.music_assistant.client.ui.compose.common.items.ProgressActions
 import io.music_assistant.client.ui.compose.common.items.ProvideClickActions
 import io.music_assistant.client.ui.compose.common.rememberToastState
 import io.music_assistant.client.ui.compose.common.viewmodel.ActionsViewModel
@@ -65,13 +47,9 @@ import io.music_assistant.client.ui.compose.nav.TopBarLayout
 import io.music_assistant.client.ui.compose.nav.TwoRowTopAppBar
 import io.music_assistant.client.ui.compose.search.SearchInput
 import musicassistantclient.composeapp.generated.resources.Res
-import musicassistantclient.composeapp.generated.resources.cd_add_playlist
 import musicassistantclient.composeapp.generated.resources.cd_toggle_view_mode
 import musicassistantclient.composeapp.generated.resources.common_back
-import musicassistantclient.composeapp.generated.resources.library_empty
-import musicassistantclient.composeapp.generated.resources.library_error
 import musicassistantclient.composeapp.generated.resources.library_quick_search
-import musicassistantclient.composeapp.generated.resources.library_search_global
 import musicassistantclient.composeapp.generated.resources.media_type_albums
 import musicassistantclient.composeapp.generated.resources.media_type_artists
 import musicassistantclient.composeapp.generated.resources.media_type_audiobooks
@@ -80,12 +58,11 @@ import musicassistantclient.composeapp.generated.resources.media_type_playlists
 import musicassistantclient.composeapp.generated.resources.media_type_podcasts
 import musicassistantclient.composeapp.generated.resources.media_type_radio
 import musicassistantclient.composeapp.generated.resources.media_type_tracks
-import musicassistantclient.composeapp.generated.resources.playlist_add_new
 import org.jetbrains.compose.resources.stringResource
 
 @Composable
-fun ItemListScreen(
-    itemListViewModel: ItemListViewModel,
+fun LibraryListScreen(
+    libraryListViewModel: LibraryListViewModel,
     contentPadding: PaddingValues,
     actionsViewModel: ActionsViewModel,
     onNavigateClick: (AppMediaItem) -> Unit,
@@ -100,70 +77,74 @@ fun ItemListScreen(
         }
     }
     LaunchedEffect(Unit) {
-        itemListViewModel.toasts.collect { toast ->
+        libraryListViewModel.toasts.collect { toast ->
             toastState.showToast(toast)
         }
     }
 
-    val state by itemListViewModel.state.collectAsStateWithLifecycle()
-    val providerOptions by itemListViewModel.providerOptions.collectAsStateWithLifecycle()
-    val genreOptions by itemListViewModel.genreOptions.collectAsStateWithLifecycle()
+    val state by libraryListViewModel.state.collectAsStateWithLifecycle()
+    val providerOptions by libraryListViewModel.providerOptions.collectAsStateWithLifecycle()
+    val genreOptions by libraryListViewModel.genreOptions.collectAsStateWithLifecycle()
 
     TopBarLayout(
         topBar = {
-            ItemListTopBar(
+            LibraryListTopBar(
                 onBack = onBack,
-                onToggleViewMode = itemListViewModel::toggleViewMode,
+                onToggleViewMode = libraryListViewModel::toggleViewMode,
                 viewMode = state.viewMode,
                 searchQuery = state.searchQuery,
-                onSearchQueryChanged = itemListViewModel::onSearchQueryChanged,
-                onSearch = itemListViewModel::onSearch,
-                onSortChanged = { itemListViewModel.onSortChanged(it) },
+                onSearchQueryChanged = libraryListViewModel::onSearchQueryChanged,
+                onSearch = libraryListViewModel::onSearch,
+                onSortChanged = { libraryListViewModel.onSortChanged(it) },
                 mediaType = state.mediaType,
                 sortOption = state.sortOption,
                 filters = state.filters,
-                onFiltersChange = itemListViewModel::setFilters,
+                onFiltersChange = libraryListViewModel::setFilters,
                 providerOptions = providerOptions,
                 genreOptions = genreOptions,
-                onLoadFilterOptions = itemListViewModel::loadFilterOptions,
+                onLoadFilterOptions = libraryListViewModel::loadFilterOptions,
             )
         },
     ) {
-        var showCreatePlaylistDialog by rememberSaveable { mutableStateOf(false) }
         ProvideClickActions(ClickContext.LIBRARY) {
             ItemList(
-                showCreatePlaylistDialog = showCreatePlaylistDialog,
-                toastState = toastState,
+                data = state.dataState,
                 onNavigateClick = onNavigateClick,
-                onGlobalSearch = onGlobalSearch,
-                searchQuery = state.searchQuery,
                 onPlayClick = { item, option, radio, _ ->
-                    itemListViewModel.onPlayClick(item, option, radio)
+                    libraryListViewModel.onPlayClick(item, option, radio)
                 },
-                onCreatePlaylistClick = { showCreatePlaylistDialog = true },
-                onLoadMore = { itemListViewModel.loadMore() },
-                onDismissCreatePlaylistDialog = { showCreatePlaylistDialog = false },
-                onCreatePlaylist = { name ->
-                    itemListViewModel.createPlaylist(name)
-                    showCreatePlaylistDialog = false
-                },
+                onCreatePlaylist = libraryListViewModel::createPlaylist,
                 playlistActions = actionsViewModel,
                 libraryActions = actionsViewModel,
                 progressActions = actionsViewModel,
                 contentPadding = contentPadding,
-                dataState = state.dataState,
                 mediaType = state.mediaType,
-                isLoadingMore = state.isLoadingMore,
-                hasMore = state.hasMore,
                 viewMode = state.viewMode,
+                hasMore = state.hasMore,
+                isLoadingMore = state.isLoadingMore,
+                onLoadMore = libraryListViewModel::loadMore,
+                onGlobalSearch = state.searchQuery.let {
+                    if (it.isNotBlank()) {
+                        { onGlobalSearch(it) }
+                    } else {
+                        null
+                    }
+                },
             )
         }
+
+        ToastHost(
+            toastState = toastState,
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(bottom = 48.dp),
+        )
     }
 }
 
 @OptIn(ExperimentalMaterial3ExpressiveApi::class)
 @Composable
-private fun ItemListTopBar(
+private fun LibraryListTopBar(
     onBack: () -> Unit,
     onToggleViewMode: () -> Unit,
     viewMode: ViewMode,
@@ -286,162 +267,5 @@ private fun ItemListTopBar(
                 }
             },
         )
-    }
-}
-
-@Composable
-private fun ItemList(
-    modifier: Modifier = Modifier,
-    showCreatePlaylistDialog: Boolean,
-    toastState: ToastState,
-    onNavigateClick: (AppMediaItem) -> Unit,
-    onGlobalSearch: (query: String) -> Unit,
-    searchQuery: String,
-    onPlayClick: PlayHandler<AppMediaItem>,
-    onCreatePlaylistClick: () -> Unit,
-    onLoadMore: () -> Unit,
-    onDismissCreatePlaylistDialog: () -> Unit,
-    onCreatePlaylist: (String) -> Unit,
-    playlistActions: PlaylistActions,
-    libraryActions: LibraryActions,
-    progressActions: ProgressActions? = null,
-    contentPadding: PaddingValues,
-    dataState: DataState<List<AppMediaItem>>,
-    mediaType: MediaType,
-    isLoadingMore: Boolean,
-    hasMore: Boolean,
-    viewMode: ViewMode,
-) {
-    Box(modifier = modifier) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .clearFocusOnScroll(),
-        ) {
-            // Content area
-            Box(modifier = Modifier.fillMaxSize()) {
-                when (dataState) {
-                    is DataState.Loading -> LoadingState()
-                    is DataState.Error -> ErrorState()
-                    is DataState.NoData -> EmptyState(searchQuery, onGlobalSearch)
-                    is DataState.Stale,
-                    is DataState.Data,
-                        -> {
-                        val items = dataState.dataOrNull.orEmpty()
-                        if (items.isEmpty()) {
-                            EmptyState(searchQuery, onGlobalSearch)
-                        } else {
-                            Column(modifier = Modifier.fillMaxSize()) {
-                                if (mediaType == MediaType.PLAYLIST) {
-                                    OutlinedButton(
-                                        modifier = Modifier
-                                            .fillMaxWidth()
-                                            .padding(horizontal = 16.dp, vertical = 8.dp),
-                                        onClick = onCreatePlaylistClick,
-                                    ) {
-                                        Icon(
-                                            TablerIcons.Plus,
-                                            contentDescription = stringResource(Res.string.cd_add_playlist),
-                                        )
-                                        Spacer(Modifier.width(4.dp))
-                                        Text(stringResource(Res.string.playlist_add_new))
-                                    }
-                                }
-
-                                val gridState = rememberLazyGridState()
-                                AdaptiveMediaGrid(
-                                    modifier = Modifier.fillMaxSize(),
-                                    items = items,
-                                    isLoadingMore = isLoadingMore,
-                                    hasMore = hasMore,
-                                    viewMode = viewMode,
-                                    onNavigateClick = onNavigateClick,
-                                    onPlayClick = onPlayClick,
-                                    onLoadMore = onLoadMore,
-                                    gridState = gridState,
-                                    playlistActions = playlistActions,
-                                    libraryActions = libraryActions,
-                                    progressActions = progressActions,
-                                    contentPadding = contentPadding,
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        // Toast host
-        ToastHost(
-            toastState = toastState,
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(bottom = 48.dp),
-        )
-
-        // Create Playlist Dialog
-        if (showCreatePlaylistDialog) {
-            CreatePlaylistDialog(
-                onDismiss = onDismissCreatePlaylistDialog,
-                onCreate = onCreatePlaylist,
-            )
-        }
-    }
-}
-
-@Composable
-private fun LoadingState() {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center,
-    ) {
-        CircularProgressIndicator()
-    }
-}
-
-@Composable
-private fun ErrorState() {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center,
-    ) {
-        Text(
-            text = stringResource(Res.string.library_error),
-            style = MaterialTheme.typography.titleMedium,
-            color = MaterialTheme.colorScheme.error,
-        )
-    }
-}
-
-@Composable
-private fun EmptyState(
-    searchQuery: String,
-    onGlobalSearch: (query: String) -> Unit,
-) {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center,
-    ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.spacedBy(16.dp),
-        ) {
-            Text(
-                text = stringResource(Res.string.library_empty),
-                style = MaterialTheme.typography.titleMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-            // Offer escalation to global search only when an actual query yielded nothing.
-            if (searchQuery.isNotBlank()) {
-                OutlinedButton(onClick = { onGlobalSearch(searchQuery) }) {
-                    Icon(
-                        Icons.Default.Search,
-                        contentDescription = null,
-                    )
-                    Spacer(Modifier.width(8.dp))
-                    Text(stringResource(Res.string.library_search_global))
-                }
-            }
-        }
     }
 }
